# MQE: Add support for label_join