--- a/pkg/streamingpromql/operators/functions/label.go
+++ b/pkg/streamingpromql/operators/functions/label.go
@@ -8,6 +8,7 @@ package functions
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/grafana/regexp"
 	"github.com/prometheus/common/model"
@@ -17,6 +18,43 @@ import (
 	"github.com/grafana/mimir/pkg/streamingpromql/types"
 )
 
+func LabelJoinFactory(dstLabelOp, separatorOp types.StringOperator, srcLabelOp []types.StringOperator) SeriesMetadataFunction {
+	return func(seriesMetadata []types.SeriesMetadata, _ *limiting.MemoryConsumptionTracker) ([]types.SeriesMetadata, error) {
+		dst := dstLabelOp.GetValue()
+		if !model.LabelName(dst).IsValid() {
+			return nil, fmt.Errorf("invalid destination label name in label_join(): %s", dst)
+		}
+		separator := separatorOp.GetValue()
+		srcLabels := make([]string, len(srcLabelOp))
+		for i, op := range srcLabelOp {
+			src := op.GetValue()
+			if !model.LabelName(src).IsValid() {
+				return nil, fmt.Errorf("invalid source label name in label_join(): %s", dst)
+			}
+			srcLabels[i] = src
+		}
+
+		lb := labels.NewBuilder(labels.EmptyLabels())
+		labelValues := make([]string, 0, len(srcLabels))
+
+		for i := range seriesMetadata {
+			labelValues = labelValues[:0]
+
+			for _, srcLabel := range srcLabels {
+				// Get returns an empty string for missing labels, so this is safe and gives the desired output
+				// where a series may be missing a source label.
+				labelValues = append(labelValues, seriesMetadata[i].Labels.Get(srcLabel))
+			}
+
+			lb.Reset(seriesMetadata[i].Labels)
+			lb.Set(dst, strings.Join(labelValues, separator))
+			seriesMetadata[i].Labels = lb.Labels()
+		}
+
+		return seriesMetadata, nil
+	}
+}
+
 func LabelReplaceFactory(dstLabelOp, replacementOp, srcLabelOp, regexOp types.StringOperator) SeriesMetadataFunction {
 	return func(seriesMetadata []types.SeriesMetadata, _ *limiting.MemoryConsumptionTracker) ([]types.SeriesMetadata, error) {
 		regexStr := regexOp.GetValue()
@@ -25,6 +63,7 @@ func LabelReplaceFactory(dstLabelOp, replacementOp, srcLabelOp, regexOp types.St
 			return nil, fmt.Errorf("invalid regular expression in label_replace(): %s", regexStr)
 		}
 		dst := dstLabelOp.GetValue()
+		// TODO(jhesketh): Use UTF-8 validation (model.LabelName(src).IsValid()) when https://github.com/prometheus/prometheus/pull/15974 is vendored in.
 		if !model.LabelNameRE.MatchString(dst) {
 			return nil, fmt.Errorf("invalid destination label name in label_replace(): %s", dst)
 		}