--- a/pkg/streamingpromql/testdata/ours-only/functions.test
+++ b/pkg/streamingpromql/testdata/ours-only/functions.test
@@ -5,3 +5,15 @@ load 6m
 # Currently prometheus does not merge series: https://github.com/prometheus/prometheus/issues/15114
 eval range from 0 to 6m step 6m label_replace(series, "idx", "replaced", "idx", ".*")
   series{label="a", idx="replaced"} 2 4
+
+clear
+
+# label_join() tests
+load 5m
+  dup{label="a", this="a"} 1.0
+  dup{label="b", this="a"} 1.0
+
+# Prometheus fails this with enableDelayedNameRemoval disabled
+# Can be tested against both once https://github.com/prometheus/prometheus/pull/15975 is merged
+eval_fail range from 0 to 10m step 5m label_join(dup, "label", "", "this")
+  expected_fail_message vector cannot contain metrics with the same labelset