--- a/pkg/streamingpromql/testdata/upstream/functions.test
+++ b/pkg/streamingpromql/testdata/upstream/functions.test
@@ -447,28 +447,24 @@ load 5m
   testmetric{src="d",src1="e",src2="f",dst="original-destination-value"} 1
 
 # label_join joins all src values in order.
-# Unsupported by streaming engine.
-# eval instant at 0m label_join(testmetric, "dst", "-", "src", "src1", "src2")
-#   testmetric{src="a",src1="b",src2="c",dst="a-b-c"} 0
-#   testmetric{src="d",src1="e",src2="f",dst="d-e-f"} 1
+eval instant at 0m label_join(testmetric, "dst", "-", "src", "src1", "src2")
+  testmetric{src="a",src1="b",src2="c",dst="a-b-c"} 0
+  testmetric{src="d",src1="e",src2="f",dst="d-e-f"} 1
 
 # label_join treats non existent src labels as empty strings.
-# Unsupported by streaming engine.
-# eval instant at 0m label_join(testmetric, "dst", "-", "src", "src3", "src1")
-#   testmetric{src="a",src1="b",src2="c",dst="a--b"} 0
-#   testmetric{src="d",src1="e",src2="f",dst="d--e"} 1
+eval instant at 0m label_join(testmetric, "dst", "-", "src", "src3", "src1")
+  testmetric{src="a",src1="b",src2="c",dst="a--b"} 0
+  testmetric{src="d",src1="e",src2="f",dst="d--e"} 1
 
 # label_join overwrites the destination label even if the resulting dst label is empty string
-# Unsupported by streaming engine.
-# eval instant at 0m label_join(testmetric, "dst", "", "emptysrc", "emptysrc1", "emptysrc2")
-#   testmetric{src="a",src1="b",src2="c"} 0
-#   testmetric{src="d",src1="e",src2="f"} 1
+eval instant at 0m label_join(testmetric, "dst", "", "emptysrc", "emptysrc1", "emptysrc2")
+  testmetric{src="a",src1="b",src2="c"} 0
+  testmetric{src="d",src1="e",src2="f"} 1
 
 # test without src label for label_join
-# Unsupported by streaming engine.
-# eval instant at 0m label_join(testmetric, "dst", ", ")
-# 	  testmetric{src="a",src1="b",src2="c"} 0
-# 	  testmetric{src="d",src1="e",src2="f"} 1
+eval instant at 0m label_join(testmetric, "dst", ", ")
+	  testmetric{src="a",src1="b",src2="c"} 0
+	  testmetric{src="d",src1="e",src2="f"} 1
 
 # test without dst label for label_join
 load 5m
@@ -476,10 +472,9 @@ load 5m
   testmetric1{src="fizz",src1="buzz",src2="fizzbuzz"} 1
 
 # label_join creates dst label if not present.
-# Unsupported by streaming engine.
-# eval instant at 0m label_join(testmetric1, "dst", ", ", "src", "src1", "src2")
-#   testmetric1{src="foo",src1="bar",src2="foobar",dst="foo, bar, foobar"} 0
-#   testmetric1{src="fizz",src1="buzz",src2="fizzbuzz",dst="fizz, buzz, fizzbuzz"} 1
+eval instant at 0m label_join(testmetric1, "dst", ", ", "src", "src1", "src2")
+  testmetric1{src="foo",src1="bar",src2="foobar",dst="foo, bar, foobar"} 0
+  testmetric1{src="fizz",src1="buzz",src2="fizzbuzz",dst="fizz, buzz, fizzbuzz"} 1
 
 clear